users are now limited to one review per spot
